backend/src/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
pydantic
pydantic_settings
python-multipart
aiosqlite